separete SubstituteVariable out from QuadraticProgram
